--- src/com/gubertmc/maze/outline/MazeOutline3d.java
@@ -1,4 +1,0 @@
-package com.gubertmc.maze.outline;
-
-public class MazeOutline3d {
-}
